requirements.txt
autopep8==1.6.0
curses-menu==0.5.0
pycodestyle==2.8.0
pyserial==3.5
toml==0.10.2
